+++ client/src/components/Home/Home.js
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import axios from 'axios';
 
 import Layout from '../Layout/Layout';
-import { Alert, CardColumns, CardDeck, Col, Row } from 'react-bootstrap';
+import { Alert, Col, Row } from 'react-bootstrap';
 
 import ClassCard from './ClassCard';
 import Filters from './Filters';
