+++ server/src/constants/CityCategories.js
@@ -1,2 +1,1 @@
-// KEEP IN SYNC WITH THE ONE IN THE SERVER
 export const CITY_CATEGORIES = ['Munich', 'Berlin', 'Dortmund', 'Hamburg', 'Dresden'];
