+++ server/src/constants/ClassCategories.js
