--- server/src/models/CityCategories.js
@@ -1,1 +1,0 @@
-export const CITY_CATEGORIES = ['Munich', 'Berlin', 'Dortmund', 'Hamburg', 'Dresden'];
